requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
einops>=0.7.0

# Transformers and LoRA
transformers>=4.36.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Data loading
tensorflow>=2.14.0
tensorflow-datasets>=4.9.0
# gcs-connector-for-pytorch>=1.0.0  # Not available on PyPI, using TFDS for GCS access

# Utilities
PyYAML>=6.0
tqdm>=4.65.0
tensorboard>=2.14.0
wandb>=0.16.0

# GCS utilities
google-cloud-storage>=2.10.0
